+++ lib/less/rails.rb
@@ -3,11 +3,13 @@ module Less
   end
 end
 
+require 'less'
 require 'rails'
 require 'tilt'
 require 'sprockets'
 require 'sprockets/railtie'
 
 require 'less/rails/version'
+require 'less/rails/helpers'
 require 'less/rails/template_handlers'
 require 'less/rails/railtie'