+++ lib/less/rails/template_handlers.rb
@@ -1,7 +1,8 @@
 module Less  
-  module Rails
-    
+  module Rails    
     class LessTemplate < Tilt::LessTemplate
+      
+      include Helpers
       
       TO_CSS_KEYS = [:compress, :optimization, :silent, :color]
       
@@ -10,7 +11,8 @@ module Less
       
       def evaluate(scope, locals, &block)
         @output ||= begin
-          parser = ::Less::Parser.new less_parser_options(scope)
+          Less.Parser['scope'] = scope
+          parser = ::Less::Parser.new config_to_less_parser_options(scope)
           engine = parser.parse(data)
           engine.to_css config_to_css_options(scope)
         end
@@ -18,24 +20,23 @@ module Less
       
       protected
       
-      def config_paths(scope)
-        config_from_rails(scope)[:paths]
+      def config_to_less_parser_options(scope)
+        paths = config_paths(scope) + scope.environment.paths
+        {:filename => eval_file, :line => line, :paths => paths}
       end
       
       def config_to_css_options(scope)
         Hash[config_from_rails(scope).each.to_a].slice *TO_CSS_KEYS
       end
       
+      def config_paths(scope)
+        config_from_rails(scope)[:paths]
+      end
+      
       def config_from_rails(scope)
         scope.environment.context_class.less_config
       end
-
-      def less_parser_options(scope)
-        paths = config_paths(scope) + scope.environment.paths
-        {:filename => eval_file, :line => line, :paths => paths}
-      end
       
-    end
-    
+    end    
   end
 end